feat(mobile-app): Updated main activity compose ui
